Add files via upload
Changed renderer to use glMapBuffer instead of glSubBufferData so the buffer does not need to get copied every time. Also saves power and increases fps.

What this does is it when you change some objects position on the screen it should ONLY update that object without having to rebuffer entire buffer object.